--- src/main/java/sproj/yolo_porting_attempts/TrainTinyYOLO.java
@@ -1,4 +1,0 @@
-package sproj.yolo_porting_attempts;
-
-public class TrainTinyYOLO {
-}
